Tighten the constraints on the ring indicator view.
The constraints allowed for the ring indicator view to grow beyond the
frame of the rose chart view. This is necessary to allow it to keep its
1:1 aspect ratio.

However, on very small sizes (or rather, certain aspect ratios), the view
would overflow the rose chart and not align with the background or
the graphs.

Increasing the priority on two constraints (top and bottom) allows it
to still overflow on the horizontal axis in case the aspect ratio is not
1:1 of the main view.